Organize test classes in fine grained folders and namespaces
