+++ Philips.CodeAnalysis.Test/Maintainability/Maintainability/AvoidSuppressMessageAttributeAnalyzerTest.cs
@@ -5,7 +5,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Philips.CodeAnalysis.Common;
 using Philips.CodeAnalysis.MaintainabilityAnalyzers.Maintainability;
 
-namespace Philips.CodeAnalysis.Test
+namespace Philips.CodeAnalysis.Test.Maintainability.Maintainability
 {
 	[TestClass]
 	public class AvoidSuppressMessageAttributeAnalyzerTest : DiagnosticVerifier
