--- Philips.CodeAnalysis.Test/NoEmptyTestMethodsDiagnosticAnalyzerTest.cs
@@ -1,68 +1,0 @@
-﻿// © 2019 Koninklijke Philips N.V. See License.md in the project root for license information.
-
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Philips.CodeAnalysis.Common;
-using Philips.CodeAnalysis.MsTestAnalyzers;
-
-namespace Philips.CodeAnalysis.Test
-{
-	[TestClass]
-	public class NoEmptyTestMethodsDiagnosticAnalyzerTest : DiagnosticVerifier
-	{
-		#region Non-Public Data Members
-
-		#endregion
-
-		#region Non-Public Properties/Methods
-
-		protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-		{
-			return new NoEmptyTestMethodsDiagnosticAnalyzer();
-		}
-
-		protected override MetadataReference[] GetMetadataReferences()
-		{
-			return new[] { MetadataReference.CreateFromFile(typeof(TimeoutAttribute).Assembly.Location) };
-		}
-		protected override (string name, string content)[] GetAdditionalSourceCode()
-		{
-			string code = @"
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-public class DerivedTestMethod : TestMethodAttribute
-{
-}
-
-";
-
-			return new[] { ("DerivedTestMethod.cs", code) };
-		}
-
-		#endregion
-
-		#region Public Interface
-
-		[DataRow("DerivedTestMethod")]
-		[DataRow("TestMethod")]
-		[DataRow("DataTestMethod")]
-		[DataRow("DataRow(null)")]
-		[DataRow("TestInitialize")]
-		[DataRow("AssemblyInitialize")]
-		[DataRow("ClassInitialize")]
-		[DataRow("TestCleanup")]
-		[DataRow("ClassCleanup")]
-		[DataRow("AssemblyCleanup")]
-		[DataTestMethod]
-		public void EmptyMethodTriggersAnalyzer(string attribute)
-		{
-			const string template = @"using Microsoft.VisualStudio.TestTools.UnitTesting;
-[TestClass] public class Foo {{ [{0}] public void Method() {{ }} }}";
-
-			VerifyCSharpDiagnostic(string.Format(template, attribute), DiagnosticResultHelper.Create(DiagnosticIds.TestMethodsMustNotBeEmpty));
-		}
-
-		#endregion
-	}
-}
